# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Studio Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/phases:
    get:
      operationId: listProjectPhases
      tags: [phases]
      summary: List phases for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of phases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Phase"
    post:
      operationId: createPhase
      tags: [phases]
      summary: Create a phase for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePhaseRequest"
      responses:
        "201":
          description: Created phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
...
  /timeblocks:
    get:
      operationId: listTimeBlocks
      tags: [timeblocks]
      summary: List time blocks
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of time blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeBlock"
    post:
      operationId: createTimeBlock
      tags: [timeblocks]
      summary: Create a time block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimeBlockRequest"
      responses:
        "201":
          description: Created time block
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeBlock"
  /timeblocks/{id}:
    put:
      operationId: updateTimeBlock
      tags: [timeblocks]
      summary: Update a time block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimeBlockRequest"
      responses:
        "200":
          description: Updated time block
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeBlock"
    delete:
      operationId: deleteTimeBlock
      tags: [timeblocks]
      summary: Delete a time block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Time block deleted
...
components:
...
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        clientId:
          type: string
          nullable: true
        clientName:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, on_hold, completed, cancelled]
        type:
          type: string
          enum: [branding, web, interior, architecture, other]
        budgetedHours:
          type: number
        loggedHours:
          type: number
        budgetAmount:
          type: number
          nullable: true
        billedAmount:
          type: number
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - type
        - budgetedHours
        - loggedHours
        - billedAmount
        - createdAt
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        clientId:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, on_hold, completed, cancelled]
        type:
          type: string
          enum: [branding, web, interior, architecture, other]
        budgetedHours:
          type: number
        budgetAmount:
          type: number
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
      required:
        - name
        - status
        - type
        - budgetedHours
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        clientId:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, on_hold, completed, cancelled]
        type:
          type: string
          enum: [branding, web, interior, architecture, other]
        budgetedHours:
          type: number
        budgetAmount:
          type: number
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
    Phase:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        budgetedHours:
          type: number
        loggedHours:
          type: number
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [upcoming, in_progress, completed]
        kickoffDate:
          type: string
          format: date
          nullable: true
        deadlineDate:
          type: string
          format: date
          nullable: true
        pageTurnDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - budgetedHours
        - loggedHours
        - status
        - createdAt
    CreatePhaseRequest:
      type: object
      properties:
        name:
          type: string
        budgetedHours:
          type: number
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        kickoffDate:
          type: string
          format: date
          nullable: true
        deadlineDate:
          type: string
          format: date
          nullable: true
        pageTurnDate:
          type: string
          format: date
          nullable: true
      required:
        - name
        - budgetedHours
...
    TimeBlock:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        projectId:
          type: string
        projectName:
          type: string
        projectColor:
          type: string
          nullable: true
        phaseId:
          type: string
          nullable: true
        phaseName:
          type: string
          nullable: true
        allocationId:
          type: string
          nullable: true
        date:
          type: string
          format: date
        hours:
          type: number
        type:
          type: string
          enum: [work, meeting, kickoff, deadline, page_turn]
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        approved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - projectId
        - projectName
        - date
        - hours
        - type
        - approved
        - createdAt
    CreateTimeBlockRequest:
      type: object
      properties:
        userId:
          type: string
        projectId:
          type: string
        phaseId:
          type: string
          nullable: true
        allocationId:
          type: string
          nullable: true
        date:
          type: string
          format: date
        hours:
          type: number
        type:
          type: string
          enum: [work, meeting, kickoff, deadline, page_turn]
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required:
        - userId
        - projectId
        - date
        - hours
        - type
    UpdateTimeBlockRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        hours:
          type: number
        type:
          type: string
          enum: [work, meeting, kickoff, deadline, page_turn]
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        approved:
          type: boolean
...