# lib/api-zod/src/generated/types/timeBlock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TimeBlockType } from "./timeBlockType";

export interface TimeBlock {
  id: string;
  userId: string;
  userName: string;
  projectId: string;
  projectName: string;
  projectColor?: string | null;
  phaseId?: string | null;
  phaseName?: string | null;
  allocationId?: string | null;
  date: Date;
  hours: number;
  type: TimeBlockType;
  title?: string | null;
  description?: string | null;
  approved: boolean;
  createdAt: Date;
}